refactor(relay): extract additional fee logic to estimation module (Part 2)
Continues incremental refactoring based on issue #1002

Current implementation:
https://github.com/ithacaxyz/relay/blob/main/src/rpc/relay.rs#L300-L800

Changes:
- Extract remaining fee calculation logic from relay.rs to estimation/fees.rs
- Move complex fee estimation methods to dedicated module
- Reduce relay.rs complexity for better maintainability

Expected: Further reduction in relay.rs complexity and better separation of concerns
